Add - timing square to instructional trials
