+++ Assets/Scripts/Timing.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e7282a5e6156881408bb796b055799be
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
