+++ Assets/Scripts/Timing/TimingSquare.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54aadddcaf4fbf24fb2ab2eba78c5150
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
