+++ Assets/Scripts/wallSystem/PlayerController.cs
@@ -46,31 +46,7 @@ namespace wallSystem
                 goalText.text = E.Get().CurrTrial.Value.Header;
                 goalText.color = Color.white;
 
-                // testing for timing settings. 
-                Debug.Log("Timing Status: " + data.DataSingleton.GetData().TimingVerification);
 
-                // set properties of timing box
-                var timingBox = GameObject.Find("TimingUnit").GetComponent<Graphic>();
-                timingBox.GetComponent<RectTransform>().sizeDelta = new Vector2(100, 100);
-
-                // if timing diagnostics are turned on then timing square will initialize with each scene. 
-                if (data.DataSingleton.GetData().TimingVerification)
-                {
-                    timingBox.enabled = true;
-                    if (data.DataSingleton.GetData().TrialInitialValue % 2 == 0)
-                    {
-                        timingBox.color = Color.black;
-                    }
-                    else
-                    {
-                        timingBox.color = Color.white;
-                    }
-                }
-                else
-                {
-                    timingBox.enabled = false;
-
-                }
             }
             catch (NullReferenceException e)
             {
